Varchar-length extended
You can find the calculations of the maximum possible length of the
strings used in these tables below. To be reasonable, I chose a power-
of-two near 10'000, which of course is 2^13 = 8192 characters, which
should be enough.

Maximum row size of a table is 65'535 bytes. Since we have UTF-8, a
character may take up to 3 byte.

Card.orm.xml: 2 integers (4 bytes each) and 2 strings.
(65'535-2*4)/3/2 = 10'921 => VARCHAR(10921) for each string is the
maximum value.

Folder.orm.xml: 1 integer and 1 string.
(65'535-4)/3 = 21'843 => VARCHAR(21843) for this string is the maximum
value.

[http://dev.mysql.com/doc/refman/5.5/en/column-count-limit.html]
[http://dev.mysql.com/doc/refman/5.5/en/storage-requirements.html]